pyproject.toml
[tool.poetry]
name = "ontv"
version = "0.1.0"
description = ""
authors = ["cacko <alex@cacko.net>"]
readme = "README.md"

[[tool.poetry.source]]
name = "cacko"
secondary = true
url = "https://pypi.cacko.net/simple/"

[tool.poetry.dependencies]
python = "^3.10"
semver = "^2.13.0"
python-dotenv = "^0.19.1"
dataclasses-json = "^0.5.6"
requests = "^2.26.0"
Faker = "^10.0.0"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]
black = "^21.9-beta.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
